% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_deg_primer.R
\name{expand_degenerate_primer}
\alias{expand_degenerate_primer}
\title{Automatically expand a primer sequence that contains degenerate bases}
\usage{
expand_degenerate_primer(input_primer = "ATGCN")
}
\arguments{
\item{input_primer}{Primer sequences}
}
\value{
primer_list_onerow Expanded primer list
}
\description{
\code{expand_degenerate_primer} Automatically expand a primer sequence that contains degenerate bases
}
\examples{
# expand_degenerate_primer("ATGCN")
}
