% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\docType{data}
\name{primer_set}
\alias{primer_set}
\title{A list of primer sets}
\format{
An object of class \code{list} of length 21.
}
\usage{
primer_set
}
\description{
A list of primer sets
}
\keyword{data}
