% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{seq_hist}
\alias{seq_hist}
\title{Visualize sequence length distribution}
\usage{
seq_hist(path_to_fasta)
}
\arguments{
\item{path_to_fasta}{Path to FASTA file.}
}
\value{
ghist Histogram as a ggplot object
}
\description{
\code{seq_hist} Visualize sequence length distribution (a wrapper function of Bio).
}
\examples{
# seq_hist(path_to_fasta)
}
