% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDoAMP_funcs.R
\name{doamp_auto}
\alias{doamp_auto}
\title{Automatically download sequences and extract potential amplicons}
\usage{
doamp_auto(
  search_query,
  F_primer,
  R_primer,
  n_retmax = 20,
  n_retidmax = n_retmax * 10,
  n_mismatch = 0,
  random_sampling = TRUE,
  random_sampling_seed = 1234,
  output_dir = "rDoAMP_Out",
  save_parameter = TRUE,
  save_stat = TRUE,
  overwrite_output_dir = FALSE
)
}
\arguments{
\item{search_query}{Character. Search query for Entrez}

\item{F_primer}{Character. Forward primer sequence}

\item{R_primer}{Character. Reverse primer sequence}

\item{n_retmax}{Numeric. The maximum number of sequences collected from Entrez}

\item{n_retidmax}{Numeric. The maximum number of IDs collected from Entrez. Default is 10. If random_sampling = TRUE, n_retidmax IDs will be collected and then n_retmax sequences will be downloaded.}

\item{n_mismatch}{Numeric. The maximum number of primer-template mismatches allowed}

\item{random_sampling}{Logical. If TURE, n_retidmax IDs collected, and then n_retmax sequences are randomly collected}

\item{random_sampling_seed}{Numeric. Random number seed for random_sampling}

\item{output_dir}{Character. Output directory name}

\item{save_parameter}{Logical. If TRUE, parameters used in the analysis saved}

\item{save_stat}{Logical. If TRUE, summary of the analysis saved}

\item{overwrite_output_dir}{Logical. If TRUE, overwrite the contents of output directory.}
}
\description{
\code{doamp_auto} Automatically download sequences and extract potential amplicons
}
