% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDoAMP_funcs.R
\name{doamp_custom}
\alias{doamp_custom}
\title{Automatically extract potential amplicons from target sequences}
\usage{
doamp_custom(
  target_fasta,
  F_primer,
  R_primer,
  n_mismatch = 0,
  output_dir = "rDoAMP_Out",
  save_parameter = TRUE,
  save_stat = TRUE,
  overwrite_output_dir = FALSE
)
}
\arguments{
\item{target_fasta}{Target sequences}

\item{F_primer}{Forward primer sequence}

\item{R_primer}{Reverse primer sequence}
}
\value{
stats summary of the output
}
\description{
\code{doamp_custom} Automatically extract potential amplicons from target sequences
}
